backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
google-generativeai
python-dotenv
requests
beautifulsoup4
httpx
pypdf
python-docx
lxml
